+++ code/src/cljs/sixsq/nuvla/ui/deployment/events.cljs
@@ -1,12 +1,15 @@
 (ns sixsq.nuvla.ui.deployment.events
   (:require
+    [clojure.set :as set]
     [clojure.string :as str]
     [re-frame.core :refer [dispatch reg-event-db reg-event-fx]]
     [sixsq.nuvla.ui.cimi-api.effects :as cimi-api-fx]
     [sixsq.nuvla.ui.deployment.spec :as spec]
     [sixsq.nuvla.ui.deployment.utils :as utils]
+    [sixsq.nuvla.ui.job.events :as job-events]
     [sixsq.nuvla.ui.main.events :as main-events]
     [sixsq.nuvla.ui.messages.events :as messages-events]
+    [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.response :as response]
     [taoensso.timbre :as log]))
 
@@ -130,7 +133,8 @@
                 ::spec/elements-per-page] :as db} :db} [_ full-text-search]]
     {:db       (-> db
                    (assoc ::spec/full-text-search full-text-search)
-                   (assoc ::spec/page 1))
+                   (assoc ::spec/page 1)
+                   (assoc ::spec/selected-set #{}))
      :dispatch [::refresh]}))
 
 
@@ -138,9 +142,9 @@
   ::set-additional-filter
   (fn [{{:keys [::spec/page
                 ::spec/elements-per-page] :as db} :db} [_ additional-filter]]
-    {:db       (-> db
-                   (assoc ::spec/additional-filter additional-filter)
-                   (assoc ::spec/page 1))
+    {:db       (assoc db ::spec/additional-filter additional-filter
+                         ::spec/page 1
+                         ::spec/selected-set #{})
      :dispatch [::refresh]}))
 
 
@@ -170,4 +174,92 @@
   (fn [{db :db} [_ state-selector]]
     {:dispatch [::get-deployments]
      :db       (assoc db ::spec/state-selector state-selector
-                         ::spec/page 1)}))
+                         ::spec/page 1
+                         ::spec/selected-set #{})}))
+
+
+(reg-event-fx
+  ::open-modal-bulk-update
+  (fn [{{:keys [::spec/select-all?
+                ::spec/selected-set] :as db} :db} [_ filter-str module-href]]
+    (cond-> {:db (assoc db ::spec/bulk-update-modal {:filter-str  filter-str
+                                                     :module-href module-href})}
+            module-href (assoc
+                          ::cimi-api-fx/get
+                          [module-href
+                           #(dispatch
+                              [:sixsq.nuvla.ui.deployment.events/set-module-versions %])]))))
+
+
+(reg-event-db
+  ::close-modal-bulk-update
+  (fn [db _]
+    (assoc db ::spec/bulk-update-modal nil)))
+
+
+(reg-event-fx
+  ::bulk-update
+  (fn [{{:keys [::spec/select-all?
+                ::spec/selected-set
+                ::spec/full-text-search
+                ::spec/additional-filter
+                ::spec/state-selector
+                ::spec/nuvlabox]} :db}]
+    (let [state      (if (= "all" state-selector) nil state-selector)
+          filter-str (if select-all?
+                       (utils/get-filter-param full-text-search additional-filter state nuvlabox)
+                       (->> selected-set
+                            (map #(str "id='" % "'"))
+                            (apply general-utils/join-or)))]
+      {::cimi-api-fx/search
+       [:deployment (cond-> {:last        0
+                             :aggregation "terms:module/id"}
+                            (not (str/blank? filter-str)) (assoc :filter filter-str))
+        #(let [buckets      (get-in % [:aggregations :terms:module/id :buckets])
+               same-module? (= (count buckets) 1)
+               module-href  (when same-module? (-> buckets first :key))]
+           (dispatch [::open-modal-bulk-update filter-str module-href]))]})))
+
+
+(reg-event-fx
+  ::bulk-update-operation
+  (fn [{{:keys [::spec/bulk-update-modal
+                ::spec/select-all?]} :db} [_ selected-module]]
+    (let [filter-str (:filter-str bulk-update-modal)]
+      {::cimi-api-fx/operation-bulk
+                 [:deployment
+                  (fn [response]
+                    (dispatch [::job-events/wait-job-to-complete
+                               {:job-id              (:location response)
+                                :on-complete         #(js/console.log "COMPLETED: " %)
+                                :on-refresh          #(js/console.log "REFRESH: " %)
+                                :refresh-interval-ms 10000}]))
+                  "bulk-update" filter-str {:module-href selected-module}]
+       :dispatch [::close-modal-bulk-update]})))
+
+
+(reg-event-db
+  ::select-id
+  (fn [{:keys [::spec/selected-set] :as db} [_ id]]
+    (let [fn (if (utils/is-selected? selected-set id) disj conj)]
+      (update db ::spec/selected-set fn id))))
+
+
+(reg-event-db
+  ::select-all-page
+  (fn [{:keys [::spec/selected-set
+               ::spec/deployments] :as db} _]
+    (let [visible-dep-ids    (utils/visible-deployment-ids deployments)
+          all-page-selected? (utils/all-page-selected? selected-set visible-dep-ids)
+          fn                 (if all-page-selected? set/difference set/union)]
+      (-> db
+          (update ::spec/selected-set fn visible-dep-ids)
+          (assoc ::spec/select-all? false)))))
+
+
+(reg-event-db
+  ::select-all
+  (fn [db]
+    (-> db
+        (update ::spec/select-all? not)
+        (assoc ::spec/selected-set #{}))))